Move error message position and add more styles
The error message was moved back inside the input wrapper but was put within another div inside the input wrapper called email-wrapper, with the email input. This resolves an issue with positioning the error message when active. More styles were added as well to bring the mobile responsive version of the webpage closer to a finished product.